\name{home}
\alias{home}
\title{home}
\description{
This collection of functions implements selected clinical prediction rules

\itemize{
\item \href{http://pubmed.gov/24222018}{New Pooled Cohort ASCVD Risk Equations} for the 2013 ACC/AHA Guidelines to Reduce Cardiovascular Risk
\item \href{http://pubmed.gov/23649456}{Mayo Clinic model for prediction of malignancy in patients with noncalcified pulmonary nodules} for the Diagnosis and Management of Lung Cancer, 3rd ed: American College of Chest Physicians Evidence-Based Clinical Practice Guidelines
}

}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\references\{
\href{http://pubmed.gov/24222018}{New Pooled Cohort ASCVD Risk Equations} for the 2013 ACC/AHA Guidelines to Reduce Cardiovascular Risk
}
\seealso{
\itemize{
\item \code{\link{pending}}.
}
}
\keyword{openrules}
\keyword{clinical prediction rules}

http://cran.r-project.org/doc/manuals/R-exts.html
