\name{pulmnodule}
\alias{pulmnodule}
\title{pulmnodule}
\description{
This function implements the following clinical prediction rule:
\itemize{
\item \href{http://pubmed.gov/23649456}{Mayo Clinic model for prediction of malignancy in patients with noncalcified pulmonary nodules} for the Diagnosis and Management of Lung Cancer, 3rd ed: American College of Chest Physicians Evidence-Based Clinical Practice Guidelines
}
}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\references\{
\href{http://pubmed.gov/23649456}{Mayo Clinic model for prediction of malignancy in patients with noncalcified pulmonary nodules} for the Diagnosis and Management of Lung Cancer, 3rd ed: American College of Chest Physicians Evidence-Based Clinical Practice Guidelines
}
\seealso{
\itemize{
\item \code{\link{pending}}.
}
}
\keyword{openRules}
\keyword{clinical prediction rules}

http://cran.r-project.org/doc/manuals/R-exts.html
