\name{home}
\alias{home}
\title{home}
\description{
This collection of functions several clinical prediction rules.
}
\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\references{
\itemize{
	\item \href{http://pubmed.gov/24222018}{New Pooled Cohort ASCVD Risk Equations} for the 2013 ACC/AHA Guidelines to Reduce Cardiovascular Risk
	}
}
\seealso{
\itemize{
	\item \code{\link{pending}}.
	}
}
\keyword{openRules}
\keyword{clinical prediction rules}

