\name{qitools}
\alias{qitools}
\title{qitools}
\description{
This collection of functions implements the following clinical prediction rules}

\itemize{
\item \url{pending}
}

}

\author{
  Robert Badgett \email{rbadgett@kumc.edu}
}
\seealso{
\itemize{
\item \code{\link{pending}}.

}
}

http://cran.r-project.org/doc/manuals/R-exts.html
